pyproject.toml
[tool.poetry]
name = "fastapi-async-challenge"
version = "0.1.0"
description = "Desafio: API Assíncrona com FastAPI"
authors = ["Seu Nome <seu@email.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
pydantic-settings = "^2.5.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
fastapi-pagination = "^0.12.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
